% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{Boston}
\alias{Boston}
\title{Boston}
\format{
A data frame with 506 rows and 14 columns:
\describe{
    \item{crim}{numeric: As for MASS dataset of the same name. }
    \item{zn}{numeric: As for MASS dataset of the same name. }
    \item{indus}{numeric: As for MASS dataset of the same name. }
    \item{chas}{integer: As for MASS dataset of the same name. }
    \item{nox}{numeric: As for MASS dataset of the same name. }
    \item{rm}{numeric: As for MASS dataset of the same name. }
    \item{age}{numeric: As for MASS dataset of the same name. }
    \item{dis}{numeric: As for MASS dataset of the same name. }
    \item{rad}{integer: As for MASS dataset of the same name. }
    \item{tax}{numeric: As for MASS dataset of the same name. }
    \item{ptratio}{numeric: As for MASS dataset of the same name. }
    \item{black}{numeric: As for MASS dataset of the same name. }
    \item{lstat}{numeric: As for MASS dataset of the same name. }
    \item{medv}{numeric: As for MASS dataset of the same name. }
}
}
\usage{
Boston
}
\description{
Taken from the MASS data sets.  See MASS::<data set> for more information
}
\keyword{datasets}
