% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{Cars93}
\alias{Cars93}
\title{Cars93}
\format{
A data frame with 93 rows and 27 columns:
\describe{
    \item{Manufacturer}{factor: As for MASS dataset of the same name. }
    \item{Model}{factor: As for MASS dataset of the same name. }
    \item{Type}{factor: As for MASS dataset of the same name. }
    \item{Min.Price}{numeric: As for MASS dataset of the same name. }
    \item{Price}{numeric: As for MASS dataset of the same name. }
    \item{Max.Price}{numeric: As for MASS dataset of the same name. }
    \item{MPG.city}{integer: As for MASS dataset of the same name. }
    \item{MPG.highway}{integer: As for MASS dataset of the same name. }
    \item{AirBags}{factor: As for MASS dataset of the same name. }
    \item{DriveTrain}{factor: As for MASS dataset of the same name. }
    \item{Cylinders}{factor: As for MASS dataset of the same name. }
    \item{EngineSize}{numeric: As for MASS dataset of the same name. }
    \item{Horsepower}{integer: As for MASS dataset of the same name. }
    \item{RPM}{integer: As for MASS dataset of the same name. }
    \item{Rev.per.mile}{integer: As for MASS dataset of the same name. }
    \item{Man.trans.avail}{factor: As for MASS dataset of the same name. }
    \item{Fuel.tank.capacity}{numeric: As for MASS dataset of the same name. }
    \item{Passengers}{integer: As for MASS dataset of the same name. }
    \item{Length}{integer: As for MASS dataset of the same name. }
    \item{Wheelbase}{integer: As for MASS dataset of the same name. }
    \item{Width}{integer: As for MASS dataset of the same name. }
    \item{Turn.circle}{integer: As for MASS dataset of the same name. }
    \item{Rear.seat.room}{numeric: As for MASS dataset of the same name. }
    \item{Luggage.room}{integer: As for MASS dataset of the same name. }
    \item{Weight}{integer: As for MASS dataset of the same name. }
    \item{Origin}{factor: As for MASS dataset of the same name. }
    \item{Make}{factor: As for MASS dataset of the same name. }
}
}
\usage{
Cars93
}
\description{
Taken from the MASS data sets.  See MASS::<data set> for more information
}
\keyword{datasets}
