% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC_BIC.R
\name{default_test}
\alias{default_test}
\alias{default_test.default}
\alias{default_test.negbin}
\alias{default_test.lmerMod}
\alias{default_test.glmerMod}
\alias{default_test.multinom}
\alias{default_test.polr}
\alias{default_test.glm}
\alias{default_test.lm}
\title{Guess the default test}
\usage{
default_test(object)

\method{default_test}{default}(object)

\method{default_test}{negbin}(object)

\method{default_test}{lmerMod}(object)

\method{default_test}{glmerMod}(object)

\method{default_test}{multinom}(object)

\method{default_test}{polr}(object)

\method{default_test}{glm}(object)

\method{default_test}{lm}(object)
}
\arguments{
\item{object}{a fitted model object accommodated by \code{\link[MASS]{dropterm}}}
}
\value{
A character string, one of \code{"F", "Chisq", or "none"}
}
\description{
Find an appropriate test to use in \code{\link[MASS]{dropterm}} if not specified
}
\examples{
fm <- glm.nb(Days ~ .^3, quine)
default_test(fm)
}
