% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_helper_functions.R
\name{.normalise}
\alias{.normalise}
\title{Normalise a vector}
\usage{
.normalise(x, location, scale)
}
\arguments{
\item{x}{A numeric vector}

\item{location}{A numeric vector of length 1}

\item{scale}{A numeric vector of length 1, usually positive}
}
\value{
A normalised vector inheriting from class "normalise"
}
\description{
Similar to base::scale() but returning a vector with class attribute.
Used for safe prediction
}
