% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_lt.R
\name{gs_orth_modified}
\alias{gs_orth_modified}
\alias{gs_orth}
\title{Gram-Schmidt orthogonalization}
\usage{
gs_orth_modified(X)

gs_orth(X)
}
\arguments{
\item{X}{a numerical matrix with ncol(X) <= nrow(X)}
}
\value{
A list with two components, Q, R, as usually defined.
}
\description{
Either classical or modified algorithms.
The modified algorithm is the more accurate.
}
\examples{
set.seed(1234)
X <- matrix(rnorm(10*7), 10)
gs_orth_modified(X)
all.equal(gs_orth(X), gs_orth_modified(X))
all.equal(gs_orth_modified(X), givens_orth(X))
}
