% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_helper_functions.R
\name{makepredictcall.normalise}
\alias{makepredictcall.normalise}
\title{Method function for safe prediction}
\usage{
\method{makepredictcall}{normalise}(var, call)
}
\arguments{
\item{var}{A numeric variable}

\item{call}{A single term from a linear model formula}
}
\value{
A call object used in safe prediction
}
\description{
This is an internal function not intended to be called directly by the user.
}
