% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_helper_functions.R
\name{mean_c}
\alias{mean_c}
\alias{var_c}
\title{Mean and variance for a circular sample}
\usage{
mean_c(theta)

var_c(theta)
}
\arguments{
\item{theta}{A vector of angles (in radians)}
}
\value{
The mean (rsp. variance) of the angle sample
}
\description{
Mean and variance for a circular sample
}
\examples{
th <- 2*base::pi*(rbeta(2000, 1.5, 1.5) - 0.5)
c(mn = mean_c(th), va = var_c(th))
rm(th)
}
