% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC_BIC.R
\name{plot.drop_term}
\alias{plot.drop_term}
\title{drop_term plot method}
\usage{
\method{plot}{drop_term}(
  x,
  ...,
  horiz = TRUE,
  las = ifelse(horiz, 1, 2),
  col = c("#DF536B", "steel blue"),
  border = c("#DF536B", "steel blue"),
  show.model = TRUE
)
}
\arguments{
\item{x}{An object of class \code{"drop_term"}}

\item{..., horiz}{arguments past on to \code{graphics::barplot}}

\item{las}{graphics parameter}

\item{show.model}{logical: should the model itself be displayed?}

\item{fill, colour}{\code{barplot} fill and border colour(s)}
}
\value{
\code{x} invisibly
}
\description{
drop_term plot method
}
\examples{
fm <- lm(medv ~ . + (rm + tax + lstat)^2 +
           I((rm - 6)^2) + I((tax - 400)^2) + I((lstat - 12)^2), Boston)
d <- drop_term(fm, k = "bic")
plot(d)
plot(d, horiz = FALSE)
}
