% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox.R
\name{print.lambda}
\alias{print.lambda}
\title{Print method for Box-Cox objects}
\usage{
\method{print}{lambda}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"box_cox"}}

\item{...}{ignored}
}
\value{
x, invisibly
}
\description{
Print method for Box-Cox objects
}
