% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{quine}
\alias{quine}
\title{quine}
\format{
A data frame with 146 rows and 5 columns:
\describe{
    \item{Eth}{factor: As for MASS dataset of the same name. }
    \item{Sex}{factor: As for MASS dataset of the same name. }
    \item{Age}{factor: As for MASS dataset of the same name. }
    \item{Lrn}{factor: As for MASS dataset of the same name. }
    \item{Days}{integer: As for MASS dataset of the same name. }
}
}
\usage{
quine
}
\description{
Taken from the MASS data sets.  See MASS::<data set> for more information
}
\keyword{datasets}
