% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASS_transfers.R, R/some_helper_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Null}
\alias{fractions}
\alias{ginv}
\alias{glm.convert}
\alias{glm.nb}
\alias{glmmPQL}
\alias{isoMDS}
\alias{kde2d}
\alias{lda}
\alias{lm.gls}
\alias{logtrans}
\alias{lqs}
\alias{negative.binomial}
\alias{polr}
\alias{qda}
\alias{rlm}
\alias{rnegbin}
\alias{sammon}
\alias{stdres}
\alias{studres}
\alias{theta.ml}
\alias{theta.mm}
\alias{theta.md}
\alias{ucv}
\alias{width.SJ}
\alias{stepAIC}
\alias{dropterm}
\alias{boxcox}
\alias{addterm}
\alias{ns}
\alias{bs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{MASS}{\code{\link[MASS]{Null}}, \code{\link[MASS]{addterm}}, \code{\link[MASS]{boxcox}}, \code{\link[MASS]{dropterm}}, \code{\link[MASS]{fractions}}, \code{\link[MASS]{ginv}}, \code{\link[MASS]{glm.convert}}, \code{\link[MASS]{glm.nb}}, \code{\link[MASS]{glmmPQL}}, \code{\link[MASS]{isoMDS}}, \code{\link[MASS]{kde2d}}, \code{\link[MASS]{lda}}, \code{\link[MASS]{lm.gls}}, \code{\link[MASS]{logtrans}}, \code{\link[MASS]{lqs}}, \code{\link[MASS]{negative.binomial}}, \code{\link[MASS]{polr}}, \code{\link[MASS]{qda}}, \code{\link[MASS]{rlm}}, \code{\link[MASS]{rnegbin}}, \code{\link[MASS]{sammon}}, \code{\link[MASS]{stdres}}, \code{\link[MASS]{stepAIC}}, \code{\link[MASS]{studres}}, \code{\link[MASS]{theta.md}}, \code{\link[MASS:theta.md]{theta.ml}}, \code{\link[MASS:theta.md]{theta.mm}}, \code{\link[MASS]{ucv}}, \code{\link[MASS]{width.SJ}}}

  \item{splines}{\code{\link[splines]{bs}}, \code{\link[splines]{ns}}}
}}

