% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoid.R
\name{unitChange}
\alias{unitChange}
\alias{cm2in}
\alias{mm2in}
\alias{in2cm}
\alias{in2mm}
\title{Unit change functions}
\usage{
cm2in(cm)

mm2in(mm)

in2cm(inch)

in2mm(inch)
}
\arguments{
\item{cm, inch, mm}{numeric vectors in the appropriate units}
}
\value{
a numeric vector of values in the new units
}
\description{
Convert imperial to metric units, and vice versa.
}
