% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.nb.vcov.R
\name{vcovx}
\alias{vcovx}
\alias{vcovx.default}
\alias{vcovx.negbin}
\title{Extended variance matrix}
\usage{
vcovx(object, ...)

\method{vcovx}{default}(object, ...)

\method{vcovx}{negbin}(object, ...)
}
\arguments{
\item{object}{A fitted mdel objeds}

\item{...}{currently ignored}
}
\value{
An extended variance matrix including parameters addition to the regression coefficients
}
\description{
An extension to the \code{\link[stats]{vcov}} function mainly to
cover the additional parameter involved in negative binomial models.
(Currently the same as \code{\link[stats]{vcov}} apart from negative binomial models.)
}
\examples{
fm <- glm.nb(Days ~ Sex/(Age + Eth*Lrn), quine)
Sigma <- vcovx(fm)
}
