% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{whiteside}
\alias{whiteside}
\title{whiteside}
\format{
A data frame with 56 rows and 3 columns:
\describe{
    \item{Insul}{factor: As for MASS dataset of the same name. }
    \item{Temp}{numeric: As for MASS dataset of the same name. }
    \item{Gas}{numeric: As for MASS dataset of the same name. }
}
}
\usage{
whiteside
}
\description{
Taken from the MASS data sets.  See MASS::<data set> for more information
}
\keyword{datasets}
