% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avoid.R
\docType{class}
\name{xy-class}
\alias{xy-class}
\title{An S4 class to represent alternavive complex, matrix or list input forms.}
\description{
An S4 class to represent alternavive complex, matrix or list input forms.
}
